refactor: 매번 로또 번호 리스트를 생성하지 않도록 수정
- 참조 URL: https://joosjuliet.github.io/flyweight_pattern/
